+++ examples/react-apollo/src/sign-up/index.tsx
@@ -1,4 +1,4 @@
-import { FaLock } from 'react-icons/fa'
+import { FaFingerprint, FaLock } from 'react-icons/fa'
 import { Link, Route, Routes } from 'react-router-dom'
 
 import { Anchor, Center, Divider, Text } from '@mantine/core'
@@ -10,6 +10,7 @@ import OAuthLinks from '../components/OauthLinks'
 
 import { EmailPassword } from './email-password'
 import { EmailPasswordless } from './email-passwordless'
+import { SecurityKeySignUp } from './security-key'
 
 const Index: React.FC = () => {
   const isAnonymous = useUserIsAnonymous()
@@ -21,8 +22,11 @@ const Index: React.FC = () => {
           <Divider my="sm" />
         </>
       )}
+      <AuthLink leftIcon={<FaFingerprint />} variant="outline" link="/sign-up/security-key">
+        Continue with a security key
+      </AuthLink>
       <AuthLink leftIcon={<FaLock />} variant="outline" link="/sign-up/email-passwordless">
-        Continue with passwordless email
+        Continue with a magic link
       </AuthLink>
       <AuthLink variant="subtle" link="/sign-up/email-password">
         Continue with email + password
@@ -33,13 +37,13 @@ const Index: React.FC = () => {
 export const SignUpPage: React.FC = () => {
   return (
     <AuthLayout
-      title="Sign up"
+      title="Sign up to the application"
       footer={
         <Center>
           <Text>
             Already have an account?{' '}
             <Anchor component={Link} to="/sign-in">
-              Log in
+              Sign in
             </Anchor>
           </Text>
         </Center>
@@ -49,6 +53,7 @@ export const SignUpPage: React.FC = () => {
         <Route path="/" element={<Index />} />
         <Route path="/email-password" element={<EmailPassword />} />
         <Route path="/email-passwordless" element={<EmailPasswordless />} />
+        <Route path="/security-key" element={<SecurityKeySignUp />} />
       </Routes>
     </AuthLayout>
   )