+++ packages/stripe-graphql-js/src/types.ts
@@ -3,7 +3,7 @@ import type Stripe from 'stripe'
 import type { CORSOptions, YogaInitialContext } from '@graphql-yoga/node'
 
 export type StripeGraphQLContext = {
-  isAllowed: (stripeCustomerId: string, context: Context) => boolean
+  isAllowed: (stripeCustomerId: string, context: Context) => boolean | Promise<boolean>
   userClaims?: UserHasuraClaims
   isAdmin: boolean
 }
@@ -12,7 +12,8 @@ export type Context = YogaInitialContext & StripeGraphQLContext
 
 export type CreateServerProps = {
   cors?: CORSOptions
-  isAllowed?: (stripeCustomerId: string, context: Context) => boolean
+  isAllowed?: (stripeCustomerId: string, context: Context) => boolean | Promise<boolean>
+  graphiql?: boolean
 }
 
 // removing Stripe.Customer from `customer` because we will never expand